+++ hdf5-types/src/lib.rs
@@ -19,7 +19,7 @@ pub mod dyn_value;
 mod h5type;
 mod string;
 
-pub use self::array::{Array, VarLenArray};
+pub use self::array::VarLenArray;
 pub use self::dyn_value::{DynValue, OwnedDynValue};
 pub use self::h5type::{
     CompoundField, CompoundType, EnumMember, EnumType, FloatSize, H5Type, IntSize, TypeDescriptor,